PCB information for sensor board and updates to README.
Changing the overall design to be a multi-layer perceptron using resonance for discrimination. Instead of a single large resonator, each piezo sensor will have an associated piezo resonator. The resonators will be driven by PAM8403 dual-channel amplifier boards. Detected signals from the piezo sensors will be routed from the raspberry pi to the ESP32 and weighted signals will be output from the esp32 to the PAM8403 input channels to drive the piezo elements on layer 2.